% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{b2utc}
\alias{b2utc}
\title{Back to UTC}
\usage{
b2utc(x)
}
\arguments{
\item{x}{a datetime vector}
}
\value{
A datetime vector in UTC format
}
\description{
\code{b2utc} changes the datetime object returned from SQL to UTC.
}
