% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_functions.R
\name{check_format}
\alias{check_format}
\title{Formatting different data outputs for occupancy analyses}
\usage{
check_format(file)
}
\arguments{
\item{file}{The file path of the data you want to read in as a character.
This file should be a summarised output from emammal,
the distributed UWIN Access DB, or the UWIN cloud DB.}
}
\value{
a data.frame of the data. The format of the data is written
  in the \code{\link{attributes}} of the object and is either
  \code{'emammal'},\code{'access'}, or \code{'uwin-app'}.
}
\description{
\code{check_format} inspects the first line of a data file to determine if
  it came from emammal, the distributed UWIN Access DB, or the UWIN cloud
  DB.
}
\examples{
\dontrun{
my_path <- "path_to_occupancy_data.csv"
my_data <- check_format(my_path)

# Check the format
attributes(my_data)$format
}
}
