% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{connect2db}
\alias{connect2db}
\title{Connect to UWIN database}
\usage{
connect2db(user = "root")
}
\arguments{
\item{user}{the name of the user you want to login as. Defaults to 'root'.}
}
\value{
If the password is correctly input \code{connect2db} will return
a \code{MySQLConnection} called \code{uwidb.} to the global environment.
}
\description{
\code{connect2db} connects to the UWIN camera trapping database. It takes
no arguments but will request for the password to the UWIN database when
a connection is attempted.
}
\examples{
\dontrun{
connect2db()
}
}
