% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_ready_functions.R
\name{data_source}
\alias{data_source}
\title{Connect to UWIN database}
\usage{
data_source(x)
}
\value{
A character that says where the data came from.
}
\description{
\code{data_source} checks the current files in the uwin-dataset
folder and let's you know if they were generated from an access
db, the UWIN web app, or if it's precleaned.
}
\details{
#' @param x The filepath to a given folder of data within a city.
}
\examples{
\dontrun{
data_source(x = "./chil/1_JA16/")
}

}
