% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsutil_functions.R
\name{gsutil_copy}
\alias{gsutil_copy}
\title{Copy images from google cloud}
\usage{
gsutil_copy(images_to_copy = NULL, output_folder = NULL, ncore = 2)
}
\arguments{
\item{images_to_copy}{The output data.frame from \code{images_of}.}

\item{ncore}{a numeric scalar. The number of cores to use to copy images
from google cloud. It defaults to 2.}

\item{output_to_folder}{A character vector that denotes the full path of
a folder to copy the images to.
If the folder does not exist it will be created. A final slash cannot
be placed at the end of the character. For example, \code{"C:/users/data"}
works but \code{"C:/users/data/"} does not.}
}
\description{
\code{gsutil_copy} takes in the object from \code{images_of} and copies
  them to a specified folder.
}
