% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsutil_functions.R
\name{gsutil_delete}
\alias{gsutil_delete}
\title{Delete images from google cloud}
\usage{
gsutil_delete(images_to_delete = NULL, ncore = 2)
}
\arguments{
\item{images_to_delete}{a data.frame with a column titled 'filepath'.
the filepath is where on the cloud these photos should be
deleted from}

\item{ncore}{a numeric scalar. The number of cores to use to copy images
from google cloud. It defaults to 2.}
}
\description{
\code{gsutil_delete} takes in the object from \code{images_of} and deletes them
from the google cloud.
}
