% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{images_of}
\alias{images_of}
\title{Images of a species at a given study area}
\usage{
images_of(species = NULL, studyArea = NULL, db = uwidb)
}
\arguments{
\item{species}{A character vector of the species to be queried. If left as
\code{NULL} then you can select the species you want in a pop up list.
Multiple species can be selected by holding CTRL when clicking on them.}

\item{studyArea}{The four letter capitalized study area abbreviation for a
city. If left as \code{NULL} you can select the study area from a pop up
list. Only one study area may be selected at a time.}

\item{db}{The MariaDB connection to the UWIN database. Defaults to 'uwidb'}
}
\value{
a data.frame with the following columns:
- filepath: the location of the image on google cloud
- locationName: the site name the image was taken
- photoName: the name of the image
- commonName: the species tagged in the images
- numIndividuals: The number of individuals of the tagged species
}
\description{
\code{images_of} queries the UWIN database for a selected species and
  study area.
}
\examples{
\dontrun{
my_images <- images_of()
}
}
