% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{MODIFY}
\alias{MODIFY}
\title{Modify the records in the UWIN database.}
\usage{
MODIFY(sql = NULL, report = FALSE, db = uwidb)
}
\arguments{
\item{sql}{A SQL statement input as a character vector to be sent to the
UWIN database.}

\item{report}{Logical. Whether or not to report the number of rows
affected. Defaults to \code{FALSE}.}

\item{db}{The MariaDB connection to the UWIN database. Defaults to 'uwidb'}
}
\description{
\code{MODIFY} can be used, for example, to update or delete records in the
  UWIN database.
}
