% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{SELECT}
\alias{SELECT}
\title{Apply a select query to the UWIN database.}
\usage{
SELECT(sql = NULL, db = uwidb)
}
\arguments{
\item{sql}{A SQL statement input as a character vector to be sent to the
UWIN database.}

\item{db}{The MariaDB connection to the UWIN database. Defaults to 'uwidb'}
}
\value{
A data.frame with the output from the SQL statement.
}
\description{
\code{SELECT} will send select SQL queries to the UWIN database and return
the output.
}
\examples{
\dontrun{
my_sql <- 'SELECT * FROM Visits;'
SELECT(my_sql)
}
}
