% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{sql_IN}
\alias{sql_IN}
\title{Concatenate SQL IN statement}
\usage{
sql_IN(x, add_quotes = TRUE)
}
\arguments{
\item{x}{A vector of values to create an IN statement with.}

\item{add_quotes}{Whether to put single quotes around values.}
}
\value{
A sql_IN statement with single quotes around the values
}
\description{
\code{sql_IN} will create an IN statement from a vector of values.
}
\examples{
my_in <- c("hello", "world")
sql_IN(my_in)

my_in2 <- c(1, 2, 3)
sql_IN(my_in2, FALSE)

}
