% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{sqlpaste}
\alias{sqlpaste}
\title{Wrapper function for paste for prettier sql statements}
\usage{
sqlpaste(...)
}
\arguments{
\item{...}{one or more `R` objects, to be converted to character vectors.}
}
\value{
A nicely formatted SQL statement.
}
\description{
\code{sqlpaste} will add a new line to the end of every paste statement.
If the last object is a sole ";", it will be tacked onto the element before it.
}
\examples{
\dontrun{
tmp_qry <- sqlpaste(
"select * from tbldf",
"where tbldf.column == 'value'"
)
response <- SELECT(tmp_qry)
}
}
